SISRP-50258 - Fixes issue on Award Card which prevents some students from seeing My Financial Summary link
